Task 009: Задание 1: контейнеризация = Moved project to booksService; = Updated paths and ts configuration; + Added and customized dockerfile for booksService
