+++ booksService/.eslintrc.js
@@ -1,9 +1,12 @@
-{
+module.exports = {
   "extends": [
     "eslint:recommended",
     "plugin:@typescript-eslint/recommended"
   ],
   "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "tsconfigRootDir": __dirname,
+  },
   "plugins": ["@typescript-eslint"],
   "rules": {
     "semi": ["error", "always"],
