+++ booksService/.gitignore
@@ -1,3 +1,4 @@
 node_modules
+dist
 .env
 public/uploads
