+++ booksService/package.json
@@ -3,8 +3,10 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "start": "nodemon --watch ./**/*.ts --exec ts-node index.ts",
-    "lint": "eslint ./**/*.ts",
+    "start": "nodemon --watch ./src/**/*.ts --exec ts-node ./src/index.ts",
+    "start:prod": "node -r tsconfig-paths/register ./dist/index.js",
+    "build": "tsc",
+    "lint": "eslint ./src/**/*.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -39,7 +41,8 @@
     "dotenv": "^16.0.3",
     "eslint": "^8.41.0",
     "nodemon": "^2.0.22",
-    "ts-node": "^10.9.1"
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.2.0"
   },
   "engines": {
     "node": "18.x",
