+++ booksService/src/models/Book.ts
